refactor: slice utils
